feat: add customizable terminators
Whiel using `cargo-get` in scripts, it is often useful to be able to
change the ending character of the output -- ensuring that a newline
does *not* get printed, adding a NUL terminator, etc.

This commit adds an option `--terminator` which changes the terminator
that is printed after the requested output from `cargo-get`.

Signed-off-by: vados <vados@vadosware.io>